[TEST] ✅ Unify MultiCompanyEasyCreation testing
