+++ energy_communities_service_invoicing/tests/__init__.py
@@ -1,5 +1,4 @@
 from . import test_multicompany_easy_creation
-from . import test_multicompany_easy_creation_with_user_profiles
 from . import test_pack_type_mixin
 from . import test_product_utils_component
 from . import test_service_invoicing_components
